Re-Add crude web UI
The UI consists of a skeleton written a long time back by someone
proficient in js and horribly thrown together code by me to adapt it to
the RESTful API. I fully expect someone having to rewrite large parts of
it but for now it does the job. In particular it is not in any way
modular, contains hard-coded classifier selections and has some
assumptions that are probably going to go away as soon as we include new
postprocessing.